Move common files to common, create benchmark-lib package
