+++ test/benchmark/reasoner/iam/LanguageFeatureTest.java
@@ -18,6 +18,9 @@
 
 package com.vaticle.typedb.core.reasoner.benchmark.iam;
 
+import com.vaticle.typedb.core.reasoner.benchmark.iam.common.Benchmark;
+import com.vaticle.typedb.core.reasoner.benchmark.iam.common.BenchmarkRunner;
+import com.vaticle.typedb.core.reasoner.benchmark.iam.common.QueryParams;
 import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
