+++ test/benchmark/reasoner/iam/common/QueryParams.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.benchmark.iam;
+package com.vaticle.typedb.core.reasoner.benchmark.iam.common;
 
 import com.vaticle.typedb.common.yaml.YAML;
 import com.vaticle.typedb.core.common.exception.TypeDBException;
@@ -27,15 +27,15 @@ import java.nio.file.Path;
 public class QueryParams {
     private static final Path paramsFile = BenchmarkRunner.RESOURCE_DIRECTORY.resolve("query_params.yml");
 
-    final String permissionEmail;
-    final String permissionObject;
-    final String permissionAction;
-    final String segregationEmail;
-    final String segregationObject;
-    final String segregationAction1;
-    final String segregationAction2;
-    final String largeNegationEmail;
-    final String basicTestObject;
+    public final String permissionEmail;
+    public final String permissionObject;
+    public final String permissionAction;
+    public final String segregationEmail;
+    public final String segregationObject;
+    public final String segregationAction1;
+    public final String segregationAction2;
+    public final String largeNegationEmail;
+    public final String basicTestObject;
 
     public QueryParams(String permissionEmail, String permissionObject, String permissionAction,
                        String segregationEmail, String segregationObject, String segregationAction1, String segregationAction2,
@@ -55,7 +55,7 @@ public class QueryParams {
         return yamlMap.get(key).asString().value();
     }
 
-    static QueryParams load() {
+    public static QueryParams load() {
         YAML.Map yamlMap = null;
         try {
             yamlMap = YAML.load(paramsFile).asMap();
